extract common dllloader class (#51)
